+++ framework/table_expr.h
@@ -7,6 +7,8 @@
 
 #include <framework/table.h>
 
+typedef function<void(table_element_ptr_t element, const table_cell_t& cell, const table_column_t* column, int index)> table_expr_drawer_t;
+
 /*! Initialize and register table expression functions */
 void table_expr_initialize();
 
@@ -19,4 +21,4 @@ void table_expr_shutdown();
  *  @param length The length of the type.
  *  @param drawer The drawer to register.
  */
-void table_expr_add_type_drawer(const char* type, size_t length, const function<void(const table_cell_t& value)>& drawer);
+void table_expr_add_type_drawer(const char* type, size_t length, const table_expr_drawer_t& drawer);